qcacmn: Change WMI API's event_buf length type to uint32_t
Change all the WMI API's event buf length param to
uint32_t to adjust to event buf length param type
in the wmi_unified_event_handler().These changes are
needed to converge on the common service ready handlers.

CRs-Fixed: 1110768

+++ wmi_unified_priv.h
@@ -933,7 +933,7 @@ QDF_STATUS (*ready_extract_mac_addr)(wmi_unified_t wmi_hdl, void *ev,
 		uint8_t *macaddr);
 QDF_STATUS (*check_and_update_fw_version)(wmi_unified_t wmi_hdl, void *ev);
 uint8_t* (*extract_dbglog_data_len)(wmi_unified_t wmi_handle, void *evt_buf,
-		uint16_t *len);
+		uint32_t *len);
 QDF_STATUS (*send_ext_resource_config)(wmi_unified_t wmi_handle,
 		wmi_host_ext_resource_config *ext_cfg);
